electron/backend/requirements.txt
Flask>=2.3
Flask-Cors>=3.1
Flask-SocketIO>=6.0
eventlet>=0.33
pyserial>=3.5
